requirements.txt
bs4==0.0.2
html2text==2024.2.26
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.3
rich==13.9.4
Scrapy==2.12.0
scrapy-proxies==0.4
SQLAlchemy==2.0.36